pyproject.toml
[tool.poetry]
name = "kazan-reds"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "3.12.0"
pydantic-settings = "2.4.0"
python-slugify = "8.0.4"
aiohttp = "3.10.5"
psycopg2-binary = "^2.9.10"
aioschedule = "^0.5.2"
aiogram-dialog = "^2.2.0"
aiogram-calendar = "^0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
